GH-57 - Rage Kick operation (#90)
As a DAO member I want to be able to transfer the remaining funds to the kicked member only after the guild kick proposal has passed. Any member of the DAO that is active (voting power) is allowed to trigger the funds transfer, this operation is known as ragekick - it forces the kicked member to redeem their proportional share of tokens that were locked after the guild kick proposal was approved. This is effectively the same thing as the kicked member calling ragequit themselves.